+++ CHANGELOG.md
@@ -9,6 +9,16 @@ x.y.z Release notes (yyyy-MM-dd)
   ([Core #7552](https://github.com/realm/realm-core/pull/7552)).
 * Improve perfomance of IN queries and chained OR equality queries for
   UUID/ObjectId types. ([.Net #3566](https://github.com/realm/realm-dotnet/issues/3566))
+  UUID/ObjectId types. ([.Net * #3566](https://github.com/realm/realm-dotnet/issues/3566))
+* Added support for updating Atlas Device Sync's base url, in case the need to roam between 
+  servers (cloud and/or edge server). This API is private and can only be imported using 
+  `@_spi(Private)` 
+   ```swift
+   @_spi(Private) import RealmSwift
+  
+   try await app.updateBaseUrl(to: "https://services.cloud.mongodb.com")
+  ```
+  ([#8486](https://github.com/realm/realm-swift/issues/8486)).
 * Enable building RealmSwift as a dynamic framework when installing via SPM, which
   lets us supply a privacy manifest. When RealmSwift is built as a static
   library you must supply your own manifest, as Xcode does not build static
@@ -24,8 +34,6 @@ x.y.z Release notes (yyyy-MM-dd)
      to "Embed & Sign" in the "Frameworks, Libraries, and Embedded Content"
      section on the General tab of your target's settings.
   ([#8561](https://github.com/realm/realm-swift/pull/8561)).
-  UUID/ObjectId types. ([.Net * #3566](https://github.com/realm/realm-dotnet/issues/3566))
-* Added support for updating Atlas Device Sync's base url, in case the need to roam between servers (cloud and/or edge server). ([#8486](https://github.com/realm/realm-swift/issues/8486))
 
 ### Fixed
 * `-[RLMUser allSessions]` did not include sessions which were currently